Generate individual plots from countsimQCReport output

Generate separate plots for all evaluation criteria using the collection of ggplot objects that can be saved when generating a countsimQC report (by setting savePlots = TRUE).

generateIndividualPlots(
  ggplotsRds,
  device = "png",
  outputDir = "./",
  nDatasets = 2
)

Arguments

ggplotsRds

The path to a .rds file generated by countsimQCReport by setting savePlots = TRUE, or the list of plots stored in this file.

device

One of "eps", "ps", "tex" (pictex), "pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only) (will be provided to the ggsave function from the ggplot2 package).

outputDir

The output directory where the plots should be generated.

nDatasets

The number of data sets that are compared in the figures. This is needed to set the size of the plots correctly.

Value

Nothing is returned, but plots are generated in the designated output directory.

Author

Charlotte Soneson

Examples

## Load example data
data(countsimExample)
if (FALSE) {
## Generate report
countsimQCReport(countsimExample, outputDir = "./",
                 outputFile = "example.html", savePlots = TRUE)
## Generate individual plots
generateIndividualPlots("example_ggplots.rds", nDatasets = 3)
}